pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "english-as-code"
version = "0.1.0"
description = "Controlled Natural Language for Excel and web ERP automation"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "T. Lee" }]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0",
    "playwright>=1.40",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
eac = "eac.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
